Refactor MidnightFetcher to use MidnightClient for block fetching and update startBlockHeight in config
